requirements.txt
flask>=3.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
flask-session>=0.8.0
bcrypt>=4.1.0
flask-wtf>=1.2.0
email-validator>=2.0.0
